requirements.txt
numpy==1.26.0
Pillow==10.1.0